+++ src/file.ts
@@ -278,7 +278,7 @@ const STORAGE_DOWNLOAD_BASE_URL = 'https://storage.googleapis.com';
  * @private
  */
 const STORAGE_UPLOAD_BASE_URL =
-  'https://www.googleapis.com/upload/storage/v1/b';
+  'https://storage.googleapis.com/upload/storage/v1/b';
 
 /**
  * @const {RegExp}
@@ -295,6 +295,7 @@ export interface FileOptions {
 
 export interface CopyOptions {
   destinationKmsKeyName?: string;
+  keepAcl?: string;
   predefinedAcl?: string;
   token?: string;
   userProject?: string;
@@ -322,6 +323,7 @@ interface CopyQuery {
   rewriteToken?: string;
   userProject?: string;
   destinationKmsKeyName?: string;
+  destinationPredefinedAcl?: string;
 }
 
 interface FileQuery {
@@ -359,6 +361,7 @@ export interface CreateReadStreamOptions {
   validation?: 'md5' | 'crc32c' | false | true;
   start?: number;
   end?: number;
+  decompress?: boolean;
 }
 
 export interface SaveOptions extends CreateWriteStreamOptions {}
@@ -829,7 +832,10 @@ class File extends ServiceObject<File> {
     });
   }
 
-  copy(destination: string | Bucket | File): Promise<CopyResponse>;
+  copy(
+    destination: string | Bucket | File,
+    options?: CopyOptions
+  ): Promise<CopyResponse>;
   copy(destination: string | Bucket | File, callback: CopyCallback): void;
   copy(
     destination: string | Bucket | File,
@@ -856,6 +862,8 @@ class File extends ServiceObject<File> {
    *     `projects/my-project/locations/location/keyRings/my-kr/cryptoKeys/my-key`,
    *     that will be used to encrypt the object. Overwrites the object
    * metadata's `kms_key_name` value, if any.
+   * @property {string} [keepAcl] This parameter is not supported and will be
+   *     removed in the next major.
    * @property {string} [predefinedAcl] Set the ACL for the new file.
    * @property {string} [token] A previously-returned `rewriteToken` from an
    *     unfinished rewrite request.
@@ -982,6 +990,11 @@ class File extends ServiceObject<File> {
       options = optionsOrCallback;
     }
 
+    if (options.hasOwnProperty('keepAcl')) {
+      // TODO: remove keepAcl from interface in next major.
+      emitWarning();
+    }
+
     options = extend(true, {}, options);
     callback = callback || util.noop;
 
@@ -1019,6 +1032,10 @@ class File extends ServiceObject<File> {
     if (options.userProject !== undefined) {
       query.userProject = options.userProject;
       delete options.userProject;
+    }
+    if (options.predefinedAcl !== undefined) {
+      query.destinationPredefinedAcl = options.predefinedAcl;
+      delete options.predefinedAcl;
     }
 
     newFile = newFile! || destBucket.file(destName);
@@ -1110,6 +1127,10 @@ class File extends ServiceObject<File> {
    *     NOTE: Byte ranges are inclusive; that is, `options.start = 0` and
    *     `options.end = 999` represent the first 1000 bytes in a file or object.
    *     NOTE: when specifying a byte range, data integrity is not available.
+   * @property {boolean} [decompress=true] Disable auto decompression of the
+   *     received data. By default this option is set to `true`.
+   *     Applicable in cases where the data was uploaded with
+   *     `gzip: true` option. See {@link File#createWriteStream}.
    */
   /**
    * Create a readable stream to read the contents of the remote file. It can be
@@ -1182,6 +1203,7 @@ class File extends ServiceObject<File> {
    *   .pipe(fs.createWriteStream('/Users/stephen/logfile.txt'));
    */
   createReadStream(options: CreateReadStreamOptions = {}): Readable {
+    options = Object.assign({decompress: true}, options);
     const rangeRequest =
       typeof options.start === 'number' || typeof options.end === 'number';
     const tailRequest = options.end! < 0;
@@ -1302,7 +1324,7 @@ class File extends ServiceObject<File> {
           throughStreams.push(validateStream);
         }
 
-        if (isCompressed) {
+        if (isCompressed && options.decompress) {
           throughStreams.push(zlib.createGunzip());
         }
 
@@ -1844,6 +1866,47 @@ class File extends ServiceObject<File> {
     });
 
     return stream as Writable;
+  }
+
+  /**
+   * Delete failed resumable upload file cache.
+   *
+   * Resumable file upload cache the config file to restart upload in case of
+   * failure. In certain scenarios, the resumable upload will not works and
+   * upload file cache needs to be deleted to upload the same file.
+   *
+   * Following are some of the scenarios.
+   *
+   * Resumable file upload failed even though the file is successfully saved
+   * on the google storage and need to clean up a resumable file cache to
+   * update the same file.
+   *
+   * Resumable file upload failed due to pre-condition
+   * (i.e generation number is not matched) and want to upload a same
+   * file with the new generation number.
+   *
+   * @example
+   * const {Storage} = require('@google-cloud/storage');
+   * const storage = new Storage();
+   * const myBucket = storage.bucket('my-bucket');
+   *
+   * const file = myBucket.file('my-file', { generation: 0 });
+   * const contents = 'This is the contents of the file.';
+   *
+   * file.save(contents, function(err) {
+   *   if (err) {
+   *     file.deleteResumableCache();
+   *   }
+   * });
+   *
+   */
+  deleteResumableCache() {
+    const uploadStream = resumableUpload.upload({
+      bucket: this.bucket.name,
+      file: this.name,
+      generation: this.generation,
+    });
+    uploadStream.deleteConfig();
   }
 
   download(options?: DownloadOptions): Promise<DownloadResponse>;
@@ -2536,7 +2599,8 @@ class File extends ServiceObject<File> {
     }
 
     const extensionHeaders = Object.assign({}, config.extensionHeaders);
-    extensionHeaders.host = 'storage.googleapis.com';
+    const fqdn = new url.URL(config.cname || STORAGE_DOWNLOAD_BASE_URL);
+    extensionHeaders.host = fqdn.host;
     if (config.method === 'POST') {
       extensionHeaders['x-goog-resumable'] = 'start';
     }
@@ -2574,7 +2638,7 @@ class File extends ServiceObject<File> {
 
       const canonicalRequest = [
         config.method,
-        config.resource,
+        config.cname ? `/${config.name}` : config.resource,
         canonicalQueryParams,
         extensionHeadersString,
         signedHeaders,
@@ -2860,7 +2924,8 @@ class File extends ServiceObject<File> {
    * location) and {@link File#delete} (from the old location). While
    * unlikely, it is possible that an error returned to your callback could be
    * triggered from either one of these API calls failing, which could leave a
-   * duplicate file lingering.
+   * duplicate file lingering. The error message will indicate what operation
+   * has failed.
    *
    * @see [Objects: copy API Documentation]{@link https://cloud.google.com/storage/docs/json_api/v1/objects/copy}
    *
@@ -2970,15 +3035,28 @@ class File extends ServiceObject<File> {
 
     callback = callback || util.noop;
 
-    this.copy(destination, options, (err, destinationFile, apiResponse) => {
+    this.copy(destination, options, (err, destinationFile, copyApiResponse) => {
       if (err) {
-        callback!(err, null, apiResponse);
+        err.message = 'file#copy failed with an error - ' + err.message;
+        callback!(err, null, copyApiResponse);
         return;
       }
 
-      this.delete(options, (err, apiResponse) => {
-        callback!(err, destinationFile, apiResponse);
-      });
+      if (
+        this.name !== destinationFile!.name ||
+        this.bucket.name !== destinationFile!.bucket.name
+      ) {
+        this.delete(options, (err, apiResponse) => {
+          if (err) {
+            err.message = 'file#delete failed with an error - ' + err.message;
+            callback!(err, destinationFile, apiResponse);
+            return;
+          }
+          callback!(null, destinationFile, copyApiResponse);
+        });
+      } else {
+        callback!(null, destinationFile, copyApiResponse);
+      }
     });
   }
 
@@ -3167,8 +3245,10 @@ class File extends ServiceObject<File> {
    * @see [Per-Object Storage Class]{@link https://cloud.google.com/storage/docs/per-object-storage-class}
    * @see [Storage Classes]{@link https://cloud.google.com/storage/docs/storage-classes}
    *
-   * @param {string} storageClass The new storage class. (`multi_regional`,
-   *     `regional`, `nearline`, `coldline`)
+   * @param {string} storageClass The new storage class. (`standard`,
+   *     `nearline`, or `coldline`)
+   *     **Note:** The storage classes `multi_regional` and `regional`
+   *     are now legacy and will be deprecated in the future.
    * @param {SetStorageClassOptions} [options] Configuration options.
    * @param {string} [options.userProject] The ID of the project which will be
    *     billed for the request.
@@ -3176,7 +3256,7 @@ class File extends ServiceObject<File> {
    * @returns {Promise<SetStorageClassResponse>}
    *
    * @example
-   * file.setStorageClass('regional', function(err, apiResponse) {
+   * file.setStorageClass('nearline', function(err, apiResponse) {
    *   if (err) {
    *     // Error handling omitted.
    *   }
@@ -3187,7 +3267,7 @@ class File extends ServiceObject<File> {
    * //-
    * // If the callback is omitted, we'll return a Promise.
    * //-
-   * file.setStorageClass('regional').then(function() {});
+   * file.setStorageClass('nearline').then(function() {});
    */
   setStorageClass(
     storageClass: string,
@@ -3413,6 +3493,17 @@ promisifyAll(File, {
     'getDate',
   ],
 });
+
+let warned = false;
+export function emitWarning() {
+  if (!warned) {
+    warned = true;
+    process.emitWarning(
+      'keepAcl parameter is not supported and will be removed in the next major',
+      'DeprecationWarning'
+    );
+  }
+}
 
 /**
  * Reference to the {@link File} class.